Fix template env variable naming
When I tried to use this template env file to create my own environment variables, I encountered the issue of non-matching variable naming between the env file and the settings.py, realized that here `STRIPE_PUBLISH_KEY` was used instead of `STRIPE_PUBLIC_KEY`. This commit changes this and fixes the faulty naming to the correct one.